Refactor Nitro config and remove unused function
