+++ packages/nuxt/src/runtime/server/utils/getGlobalContextItem.ts
@@ -1,0 +1,11 @@
+import { camelCase } from 'scule'
+import type { H3Event } from 'h3'
+
+export function getGlobalContextItem(this: {
+  projectName: string
+  moduleName: string
+  event?: H3Event
+}) {
+  const mergedProjectName = camelCase(`${this.moduleName}-${this.projectName}`)
+  return this.event?.context.globalModuleContext[mergedProjectName]
+}
